+++ src/components/flyout/index.d.ts
@@ -10,6 +10,10 @@ declare module '@elastic/eui' {
      * Locks the mouse / keyboard focus to within the flyout
      */
     ownFocus?: boolean;
+    /**
+     * Specify an aria-label for the close button of the flyout.
+     */
+    closeButtonAriaLabel?: string;
   }
 
   export const EuiFlyout: React.SFC<